requirements.txt
flask
openai
slack_sdk
google-api-python-client
python-dotenv
langgraph
google-auth
google-auth-oauthlib
google-auth-httplib2
